requirements.txt
aiodns==2.0.0
aiofile==3.5.0
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
boto3==1.17.32
botocore==1.20.32
caio==0.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
geographiclib==1.50
geopy==2.1.0
gunicorn==20.0.4
haversine==2.3.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
pandas==1.1.5
Pillow==8.1.2
psycopg2==2.8.6
pycares==3.1.1
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.0.1
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.0.31
whitenoise==5.2.0
yarl==1.6.3
zipp==3.1.0
